Protect against a possible undefined chrome.windows.Window.id value
Due to a recent change in the DefinitelyTyped project [1], the `id`
property of the chrome.windows.Window interface can be undefined,
which is breaking the build for shieldsPanelReducer.ts and
rewards_panel_reducer.ts unless we add some checks to avoid
using invalid values.

This means that we can't upgrade the @types/chrome node module to
newer versions that would include that change, so this CL adds some
checks to Brave's .ts files to account for those undefined values.

[1] https://github.com/DefinitelyTyped/DefinitelyTyped/commit/ad4b187c